任务支持更换绑定执行器，方便任务分组转移和管理；

+++ xxl-job-admin/src/main/java/com/xxl/job/admin/service/impl/XxlJobServiceImpl.java
@@ -164,13 +164,19 @@ public class XxlJobServiceImpl implements XxlJobService {
 			jobInfo.setChildJobId(StringUtils.join(childJobIds, ","));
 		}
 
+		// group valid
+		XxlJobGroup jobGroup = xxlJobGroupDao.load(jobInfo.getJobGroup());
+		if (jobGroup == null) {
+			return new ReturnT<String>(ReturnT.FAIL_CODE, (I18nUtil.getString("jobinfo_field_jobgroup")+I18nUtil.getString("system_unvalid")) );
+		}
+
 		// stage job info
 		XxlJobInfo exists_jobInfo = xxlJobInfoDao.loadById(jobInfo.getId());
 		if (exists_jobInfo == null) {
 			return new ReturnT<String>(ReturnT.FAIL_CODE, (I18nUtil.getString("jobinfo_field_id")+I18nUtil.getString("system_not_found")) );
 		}
-		//String old_cron = exists_jobInfo.getJobCron();
 
+		exists_jobInfo.setJobGroup(jobInfo.getJobGroup());
 		exists_jobInfo.setJobCron(jobInfo.getJobCron());
 		exists_jobInfo.setJobDesc(jobInfo.getJobDesc());
 		exists_jobInfo.setAuthor(jobInfo.getAuthor());
@@ -186,10 +192,9 @@ public class XxlJobServiceImpl implements XxlJobService {
 
 
 		// update quartz-cron if started
-		String qz_group = String.valueOf(exists_jobInfo.getJobGroup());
-		String qz_name = String.valueOf(exists_jobInfo.getId());
         try {
-            XxlJobDynamicScheduler.updateJobCron(qz_group, qz_name, exists_jobInfo.getJobCron());
+			String qz_name = String.valueOf(exists_jobInfo.getId());
+            XxlJobDynamicScheduler.updateJobCron(qz_name, exists_jobInfo.getJobCron());
         } catch (SchedulerException e) {
             logger.error(e.getMessage(), e);
 			return ReturnT.FAIL;
@@ -201,12 +206,14 @@ public class XxlJobServiceImpl implements XxlJobService {
 	@Override
 	public ReturnT<String> remove(int id) {
 		XxlJobInfo xxlJobInfo = xxlJobInfoDao.loadById(id);
-        String group = String.valueOf(xxlJobInfo.getJobGroup());
-        String name = String.valueOf(xxlJobInfo.getId());
+		if (xxlJobInfo == null) {
+			return ReturnT.SUCCESS;
+		}
+		String name = String.valueOf(xxlJobInfo.getId());
 
 		try {
 			// unbind quartz
-			XxlJobDynamicScheduler.removeJob(name, group);
+			XxlJobDynamicScheduler.removeJob(name);
 
 			xxlJobInfoDao.delete(id);
 			xxlJobLogDao.delete(id);
@@ -222,12 +229,11 @@ public class XxlJobServiceImpl implements XxlJobService {
 	@Override
 	public ReturnT<String> start(int id) {
 		XxlJobInfo xxlJobInfo = xxlJobInfoDao.loadById(id);
-		String group = String.valueOf(xxlJobInfo.getJobGroup());
 		String name = String.valueOf(xxlJobInfo.getId());
 		String cronExpression = xxlJobInfo.getJobCron();
 
 		try {
-			boolean ret = XxlJobDynamicScheduler.addJob(name, group, cronExpression);
+			boolean ret = XxlJobDynamicScheduler.addJob(name, cronExpression);
 			return ret?ReturnT.SUCCESS:ReturnT.FAIL;
 		} catch (SchedulerException e) {
 			logger.error(e.getMessage(), e);
@@ -238,12 +244,11 @@ public class XxlJobServiceImpl implements XxlJobService {
 	@Override
 	public ReturnT<String> stop(int id) {
         XxlJobInfo xxlJobInfo = xxlJobInfoDao.loadById(id);
-        String group = String.valueOf(xxlJobInfo.getJobGroup());
         String name = String.valueOf(xxlJobInfo.getId());
 
 		try {
 			// bind quartz
-            boolean ret = XxlJobDynamicScheduler.removeJob(name, group);
+            boolean ret = XxlJobDynamicScheduler.removeJob(name);
             return ret?ReturnT.SUCCESS:ReturnT.FAIL;
 		} catch (SchedulerException e) {
 			logger.error(e.getMessage(), e);
